requirements.txt

python-dotenv
PIL
groq
streamlit
Pillow
